+++ BRPeer.h
@@ -194,7 +194,7 @@ void BRPeerRerequestBlocks(BRPeer *peer, UInt256 fromBlock);
 inline static size_t BRPeerHash(const void *peer)
 {
     // (address xor port)*FNV_PRIME
-    return (((const BRPeer *)peer)->address.u32[3] ^ ((const BRPeer *)peer)->port)*0x01000193;
+    return (size_t)((((const BRPeer *)peer)->address.u32[3] ^ ((const BRPeer *)peer)->port)*0x01000193);
 }
 
 // true if a and b have the same address and port
