+++ BRWallet.h
@@ -48,7 +48,8 @@ inline static int BRUTXOEq(const void *utxo, const void *otherUtxo)
 
 inline static size_t BRUTXOHash(const void *utxo)
 {
-    return (((const BRUTXO *)utxo)->hash.u32[0] ^ ((const BRUTXO *)utxo)->n)*0x01000193; // (hash xor n)*FNV_PRIME
+    // (hash xor n)*FNV_PRIME
+    return (size_t)((((const BRUTXO *)utxo)->hash.u32[0] ^ ((const BRUTXO *)utxo)->n)*0x01000193);
 }
 
 typedef struct BRWalletStruct BRWallet;
